+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   déclenche
 - Ajout de `@stylistic/eslint-plugin` et mise à jour des règles eslint
 - Ajout d'un prop dans `ClassicPopover` pour modifier le z-index
+- Ajout d'un slot `option` pour `ClassicSelect`
 - Ajustement de `OctopusMultiselect` pour fonctionner correctement avec des
   objets identiques avec références différentes